Fixes #2 - more testing needed. Possibily breaking change.
To address #2, we actually change how the function works - this will
probably break some uses of the tag_tales(), but that is OK because 1)
the function is still in development, so those uses should be
tentative and 2) these changes bring tag_tales() back to the original
use.

tag_tales() was expanded to allow the user to pass in a full vector
rather than a column name to address a bug that was
happening. However, we now speculate that bug was this same one, but
allowing the full column to be passed in obscured this bug. Since the
original intent was to pass a column name, and passing in the column
name is a simplier implementation, we went with this option.

The second change is to use difftime rather than diff to ensure the
resulting differences were in seconds, rather than possibly minutes,
hours, etc. We currently have the unit hardcoded into the function,
but this could be make user adjustable. More thought needed.

tag_tales() should still be considered experimental for now - we need
to build in more tests to check for correctness, especially for the
splitting of visits into sub-visits.