Edits to tag_tales for addressing #2 and additions to tests/test_tag_tales.R
Changed `||` back to `&&` in checking Station_col input; removed !is.character(Station_col) since that gets coerced in another test in that same line.  Added a test for a numeric of length 1, as well as a test for identical outputs when the input data frame is ordered by date time first or not.