Many changes - possible refactor of list functions.
This commit creates a large function for creating the data payload for
all endpoints using the structure programmatically extracted from the
docs.md file and edited by hand.

This function needs extensive testing to ensure it creates a valid
data structure for all end points. However, it provides a possible
simplification of the overall package by providing 1) a complex
function for creating all possible payloads for all end points, and 2)
a generic method for sending these data payloads to the server and
receiving the response.

Lots more needed, but this provides the basic framework.