Adds feature to return the JSON payload rather than sending request.
This feature is useful for verifying beharior outside of R, e.g. via
submitting requests from curl.

The implementation is a bit clunky, but is probably sufficient given
the lack of resources to implement a better solution (e.g. a closure
to store the json a la RCurls header handler).